First version, focused on GUI
First aproach where login, profile and subject's screen have been
designed.
Data used as example has been stored as arraylists (the database has a
bunch of users who have a bunch of subjects which have a bunch of
tests). It is just to focus on the design of the graphic UI.